Move CalculateCrystal
