+++ .Lib9c.Tests/CrystalCalculatorTest.cs
@@ -1,20 +1,23 @@
 namespace Lib9c.Tests
 {
+    using System.Collections;
     using System.Collections.Generic;
     using Nekoyume.Helper;
+    using Nekoyume.Model.Item;
     using Nekoyume.TableData;
     using Xunit;
 
     public class CrystalCalculatorTest
     {
+        private readonly TableSheets _tableSheets;
         private readonly EquipmentItemRecipeSheet _equipmentItemRecipeSheet;
         private readonly WorldUnlockSheet _worldUnlockSheet;
 
         public CrystalCalculatorTest()
         {
-            var tableSheets = new TableSheets(TableSheetsImporter.ImportSheets());
-            _equipmentItemRecipeSheet = tableSheets.EquipmentItemRecipeSheet;
-            _worldUnlockSheet = tableSheets.WorldUnlockSheet;
+            _tableSheets = new TableSheets(TableSheetsImporter.ImportSheets());
+            _equipmentItemRecipeSheet = _tableSheets.EquipmentItemRecipeSheet;
+            _worldUnlockSheet = _tableSheets.WorldUnlockSheet;
         }
 
         [Theory]
@@ -31,6 +34,61 @@ namespace Lib9c.Tests
         public void CalculateWorldUnlockCost(IEnumerable<int> worldIds, int expected)
         {
             Assert.Equal(expected * CrystalCalculator.CRYSTAL, CrystalCalculator.CalculateWorldUnlockCost(worldIds, _worldUnlockSheet));
+        }
+
+        [Theory]
+        [ClassData(typeof(CalculateCrystalData))]
+        public void CalculateCrystal((int equipmentId, int level)[] equipmentInfos, int monsterCollectionLevel, int expected)
+        {
+            var equipmentList = new List<Equipment>();
+            foreach (var (equipmentId, level) in equipmentInfos)
+            {
+                var row = _tableSheets.EquipmentItemSheet[equipmentId];
+                var equipment =
+                    ItemFactory.CreateItemUsable(row, default, 0, level);
+                equipmentList.Add((Equipment)equipment);
+            }
+
+            Assert.Equal(
+                expected * CrystalCalculator.CRYSTAL,
+                CrystalCalculator.CalculateCrystal(
+                    equipmentList,
+                    _tableSheets.CrystalEquipmentGrindingSheet,
+                    monsterCollectionLevel,
+                    _tableSheets.CrystalMonsterCollectionMultiplierSheet
+                )
+            );
+        }
+
+        private class CalculateCrystalData : IEnumerable<object[]>
+        {
+            private readonly List<object[]> _data = new List<object[]>
+            {
+                new object[]
+                {
+                    new[]
+                    {
+                        (10100000, 0),
+                        (10100000, 2),
+                    },
+                    0,
+                    300,
+                },
+                new object[]
+                {
+                    new[]
+                    {
+                        (10100000, 1),
+                        (10100000, 2),
+                    },
+                    3,
+                    390,
+                },
+            };
+
+            public IEnumerator<object[]> GetEnumerator() => _data.GetEnumerator();
+
+            IEnumerator IEnumerable.GetEnumerator() => _data.GetEnumerator();
         }
     }
 }
