+++ Lib9c/Helper/CrystalCalculator.cs
@@ -1,7 +1,11 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using Libplanet.Assets;
+using Nekoyume.Model.Item;
 using Nekoyume.TableData;
+using Nekoyume.TableData.Crystal;
 
 namespace Nekoyume.Helper
 {
@@ -25,6 +29,27 @@ namespace Nekoyume.Helper
             return worldIds
                 .Select(id => worldUnlockSheet.OrderedList.First(r => r.WorldIdToUnlock == id))
                 .Aggregate(cost, (current, row) => current + row.CRYSTAL * CRYSTAL);
+        }
+
+        public static FungibleAssetValue CalculateCrystal(
+            IEnumerable<Equipment> equipmentList,
+            CrystalEquipmentGrindingSheet crystalEquipmentGrindingSheet,
+            int monsterCollectionLevel,
+            CrystalMonsterCollectionMultiplierSheet crystalMonsterCollectionMultiplierSheet
+        )
+        {
+            Currency currency = new Currency("CRYSTAL", 18, minters: null);
+            FungibleAssetValue crystal = 0 * currency;
+            foreach (var equipment in equipmentList)
+            {
+                CrystalEquipmentGrindingSheet.Row grindingRow = crystalEquipmentGrindingSheet[equipment.Id];
+                int level = Math.Max(0, equipment.level - 1);
+                crystal += BigInteger.Pow(2, level) * grindingRow.CRYSTAL * currency;
+            }
+            CrystalMonsterCollectionMultiplierSheet.Row multiplierRow =
+                crystalMonsterCollectionMultiplierSheet[monsterCollectionLevel];
+            var extra = crystal.DivRem(100, out _) * multiplierRow.Multiplier;
+            return crystal + extra;
         }
     }
 }
